pdp initializer
